refactor(controls): extract panels and canonical RendererSwitcher path (#914)
Add ParamSlidersPanel, RendererBackendPanel, and SlotStackPanel; route
production Controls through RendererBackendPanel. Mark legacy toggles deprecated.
